+++ jest.config.js
@@ -1,11 +1,3 @@
 module.exports = {
-  testMatch: ['**/+(*.)+(spec|test).+(ts|js)?(x)'],
-  transform: {
-    '^.+\\.(ts|js|html)$': 'ts-jest',
-  },
-  resolver: '@nrwl/jest/plugins/resolver',
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageReporters: ['html'],
-  projects: '<rootDir>/libs/ui/u-users',
-  projects: '<rootDir>/libs/ui/u-products',
+  projects: ['<rootDir>/apps/innovatech-ui', '<rootDir>/libs/ui/u-auth', '<rootDir>/libs/ui/u-tests'],
 };