+++ libs/api/a-state/src/lib/guarantees/entities/index.ts
@@ -3,4 +3,3 @@ export * from './client.entity';
 export * from './guarantee.entity';
 export * from './moral-person.entity';
 export * from './physical-person.entity';
-export * from './vechicle.entity';