+++ libs/api/a-vehicles/src/lib/dto/get-vehicles.dto.ts
@@ -1,0 +1,39 @@
+import { convertStringToNumberArray } from '@ivt/c-utils';
+import { Transform } from 'class-transformer';
+import { IsArray, IsNotEmpty, IsNumberString, IsOptional } from 'class-validator';
+
+export class GetVehiclesDto {
+  @IsOptional()
+  @IsNotEmpty()
+  @IsNumberString()
+  offset;
+
+  @IsOptional()
+  @IsNotEmpty()
+  @IsNumberString()
+  limit;
+
+  @IsOptional()
+  @IsNotEmpty()
+  text: string;
+
+  @IsOptional()
+  @IsNotEmpty()
+  sort: string;
+
+  @IsOptional()
+  @IsNotEmpty()
+  direction: string;
+
+  @IsOptional()
+  @IsNotEmpty()
+  @IsArray()
+  @Transform(({ obj, key }) => convertStringToNumberArray(obj[key]))
+  companyIds: number[];
+
+  @IsOptional()
+  @IsNotEmpty()
+  @IsArray()
+  @Transform(({ obj, key }) => convertStringToNumberArray(obj[key]))
+  groupIds: number[];
+}