+++ libs/ui/u-auth/src/lib/auth-routing.module.ts
@@ -7,7 +7,6 @@ import { ResetPasswordFormContainerComponent } from './containers/reset-password
 import { SetPasswordFormContainerComponent } from './containers/set-password-form-container/set-password-form-container.component';
 import { SignInFormContainerComponent } from './containers/sign-in-form-container/sign-in-form-container.component';
 import { SetPasswordGuard } from './guards/set-password.guard';
-import { SetPasswordResolverService } from './resolvers/set-password-resolver.service';
 
 export const routes: Routes = [
   {
@@ -26,7 +25,6 @@ export const routes: Routes = [
         path: 'set-password/:passwordToken/:userId',
         component: SetPasswordFormContainerComponent,
         canActivate: [SetPasswordGuard],
-        resolve: { resolvedRoute: SetPasswordResolverService },
       },
       {
         path: 'expired-token',