+++ libs/ui/u-auth/src/lib/guards/auth.guard.spec.ts
@@ -1,4 +1,7 @@
 import { TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+import { AuthService } from '@ivt/u-state';
+import { AuthServiceMock } from '@ivt/u-tests';
 
 import { AuthGuard } from './auth.guard';
 
@@ -6,7 +9,10 @@ describe('AuthGuard', () => {
   let guard: AuthGuard;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [RouterTestingModule],
+      providers: [{ provide: AuthService, useClass: AuthServiceMock }],
+    });
     guard = TestBed.inject(AuthGuard);
   });